requirements.txt
logging==0.4.9.6
pyserial==3.5
